Flesh out project service

+++ metalpet-model/src/main/java/org/pemacy/metalpet/model/input/StandardInputType.java
@@ -14,6 +14,6 @@ public enum StandardInputType implements InputType {
 	INTEGER,
 
 	@JsonProperty("decimal")
-	DECIMAL,
+	DOUBLE,
 
 }
